KVM: arm64: Make kvm_call_hyp() a function call at hyp
kvm_call_hyp() has some logic to issue a function call or a hypercall
depending the EL at which the kernel is running. However, all the code
compiled under __KVM_NVHE_HYPERVISOR__ is guaranteed to run only at EL2,
and in this case a simple function call is needed.

As such, add some ifdefery to kvm_host.h to symplify kvm_call_hyp() when
it is used in code that is compiled into .hyp.text, as will soon be the
case for the page-table allocator.

Signed-off-by: Quentin Perret <qperret@google.com>